+++ collector/log.go
@@ -138,7 +138,7 @@ func logParse(out string, logger log.Logger) (LogMetric, error) {
 			} else {
 				val, err := parseFloat(record[i])
 				if err != nil {
-					level.Error(logger).Log("msg", "Error parsing value", "key", k, "value", record[i], "err", err)
+					level.Error(logger).Log("msg", "Error parsing value", "key", k, "value", record[i], "record", strings.Join(record, ","), "err", err)
 					return LogMetric{}, err
 				}
 				if strings.HasSuffix(k, "_MB") {