+++ collector/occupancy.go
@@ -165,7 +165,7 @@ func occupancyParse(out string, logger log.Logger) ([]OccupancyMetric, error) {
 			} else {
 				val, err := parseFloat(record[i])
 				if err != nil {
-					level.Error(logger).Log("msg", "Error parsing value", "key", k, "value", record[i], "err", err)
+					level.Error(logger).Log("msg", "Error parsing value", "key", k, "value", record[i], "record", strings.Join(record, ","), "err", err)
 					return nil, err
 				}
 				if strings.HasSuffix(k, "_MB") {